### 2018-06-04 #763 @danielskatz
Hi @jfjlaros

1 - There's an error compiling your paper.md file which seems to be that there is no bibliography file included.  You need to add a bibliography to the directory that contains paper.md (docs in this case.)

2 - Can you suggest some potential reviewers, perhaps from [this list of people](https://bit.ly/joss-reviewers) who have already agreed to review for JOSS and may be suitable for this submission?
