### 2018-06-05 4e9459d
Added bibliography.
