+++ docs/paper.bib
@@ -1,0 +1,34 @@
+@misc{structPython,
+  title={struct — Interpret strings as packed binary data.},
+  author={Python Software Foundation},
+  url={https://docs.python.org/2/library/struct.html},
+  year={2001-2018}
+}
+
+@misc{structJs,
+  title={python-struct - Packs/Unpacks/Measures structs according to Python's struct format.},
+  author={Daniel Cohen Gindi},
+  url={https://github.com/danielgindi/node-python-struct},
+  year={2013}
+}
+
+@misc{yamlPython,
+  title={PyYAML - The next generation YAML parser and emitter for Python.},
+  author={Kirill Simonov},
+  url={https://github.com/yaml/pyyaml},
+  year={2011-2015}
+}
+
+@misc{yamlJs,
+  title={JS-YAML - YAML 1.2 parser / writer for JavaScript.},
+  author={Vitaly Puzrin},
+  url={https://github.com/nodeca/js-yaml},
+  year={2011-2015}
+}
+
+@misc{ivonvJs,
+  title={iconv-lite - Pure JS character encoding conversion.},
+  author={Alexander Shtuchkin},
+  url={https://github.com/ashtuchkin/iconv-lite},
+  year={2011}
+}
