### 2018-06-05 #763 @jfjlaros
Dear @danielskatz

Thank you for pointing this out. I have added a bibliography for the essential dependencies.

As for reviewers, I think hainesr (standards / protocols), smattr (compilers) and Funnelfiasco (CLI tools) may be interested.
