### 2018-06-05 8cd6dc8
fixes to bib
