+++ docs/paper.bib
@@ -1,33 +1,33 @@
 @misc{structPython,
   title={struct — Interpret strings as packed binary data.},
-  author={Python Software Foundation},
+  author={{Python Software Foundation}},
   url={https://docs.python.org/2/library/struct.html},
   year={2001-2018}
 }
 
 @misc{structJs,
-  title={python-struct - Packs/Unpacks/Measures structs according to Python's struct format.},
+  title={python-struct - Packs/Unpacks/Measures structs according to {P}ython's struct format.},
   author={Daniel Cohen Gindi},
   url={https://github.com/danielgindi/node-python-struct},
   year={2013}
 }
 
 @misc{yamlPython,
-  title={PyYAML - The next generation YAML parser and emitter for Python.},
+  title={PyYAML - The next generation {YAML} parser and emitter for {P}ython.},
   author={Kirill Simonov},
   url={https://github.com/yaml/pyyaml},
   year={2011-2015}
 }
 
 @misc{yamlJs,
-  title={JS-YAML - YAML 1.2 parser / writer for JavaScript.},
+  title={{JS-YAML} - {YAML} 1.2 parser / writer for Java{S}cript.},
   author={Vitaly Puzrin},
   url={https://github.com/nodeca/js-yaml},
   year={2011-2015}
 }
 
-@misc{ivonvJs,
-  title={iconv-lite - Pure JS character encoding conversion.},
+@misc{iconvJs,
+  title={iconv-lite - Pure {JS} character encoding conversion.},
   author={Alexander Shtuchkin},
   url={https://github.com/ashtuchkin/iconv-lite},
   year={2011}
