### 2018-06-05 #763 @danielskatz
@jfjlaros - Please merge https://github.com/jfjlaros/bin-parser/pull/7 to fix the remaining bib issues
