### 2018-06-05 #763 @danielskatz
@hainesr , @smattr, @Funnelfiasco - would one of you be willing to review this JOSS submission?
