+++ docs/paper.bib
@@ -20,7 +20,7 @@
 }
 
 @misc{yamlJs,
-  title={{JS-YAML} - {YAML} 1.2 parser / writer for Java{S}cript.},
+  title={{JS-YAML} - {YAML} 1.2 parser / writer for {J}ava{S}cript.},
   author={Vitaly Puzrin},
   url={https://github.com/nodeca/js-yaml},
   year={2011-2015}
