### 2018-06-05 #763 @jfjlaros
@danielskatz - Your fixes were merged and I fixed one additional capitalisation issue. Thank you.
