### 2018-06-05 d33c3a8
Removed unused line from example.
