### 2018-06-06 #763 @Smattr
I'm OK to review as well. I don't have a strong JS background, but I'm solid on Python and reverse engineering. Were you planning to pick just one or did you want multiple reviewers?
