### 2018-06-06 #763 @danielskatz
I would be pleased if we could have two reviews - I'm generally trying to do this for JOSS papers, unless I'm really an expert in the software myself, which is not the case here.
