### 2018-06-06 #766 @whedon
[REVIEW]: General binary file parser.
**Submitting author:** @jfjlaros (<a href="http://orcid.org/0000-0002-8715-7371">J.F.J. Laros</a>)
**Repository:** <a href="https://github.com/jfjlaros/bin-parser" target ="_blank">https://github.com/jfjlaros/bin-parser</a>
**Version:** 0.0.20
