### 2018-06-07 #766 @hainesr
Hi @danielskatz, for the version check, does there need to be an actual tagged release and corresponding GitHub release (i.e. here: https://github.com/jfjlaros/bin-parser/releases)?
