### 2018-06-09 jfjlaros/bin-parser#8 @Smattr
Python 3 compatibility
It looks like the Python implementation is Python2-only. E.g. [indexing the result of `map`](https://github.com/jfjlaros/bin-parser/blob/master/python/bin_parser.py#L157) will not work in Python 3 where `map` returns an iterator. Was the intention for this code to be compatible across both 2 and 3?

Related to https://github.com/openjournals/joss-reviews/issues/766.
