### 2018-06-09 jfjlaros/bin-parser#9 @Smattr
compare_yaml.py referred to as compare_yaml in docs
**Describe the bug**
[The documentation](http://bin-parser.readthedocs.io/en/latest/extras.html#compare-yaml) describes a script `compare_yaml` that is used in the test script `extras/test.sh`. This script is actually `extras/compare_yaml.py` and I had to write a wrapper around it to get `test.sh` to run correctly. Is this accidental or am I perhaps running the test script incorrectly?

**To Reproduce**
1. Run `bash extras/test.sh` as described in [the docs](http://bin-parser.readthedocs.io/en/latest/extras.html#test-sh).
2. Observe failures from being unable to find `compare_yaml`.

**Expected behavior**
The test script should run correctly to completion.

```
$ lsb_release -a
Distributor ID: Ubuntu
Description:    Ubuntu 16.04.4 LTS
Release:        16.04
Codename:       xenial
```

**Additional context**
The same issue affects `make_skeleton` in the docs.

Related to https://github.com/openjournals/joss-reviews/issues/766.
