### 2018-06-09 jfjlaros/bin-parser#10 @Smattr
Undocumented trim option
**Describe the bug**
Both the [Python sources](https://github.com/jfjlaros/bin-parser/blob/master/python/bin_parser.py#L43) and [JavaScript sources](https://github.com/jfjlaros/bin-parser/blob/master/javascript/index.js#L151) support an option called `trim`. The [overrides section of the docs](http://bin-parser.readthedocs.io/en/latest/types.html#overrides) omits this option.

**Expected behavior**
The `trim` option documented along with `size` and friends.

**Additional context**
Related to https://github.com/openjournals/joss-reviews/issues/766.
