### 2018-06-09 jfjlaros/bin-parser#12 @Smattr
Encoding flexibility?
This is very nice work, @jfjlaros, and immediately made me think of how I would use this for decoding ELF-like formats and protocol packets. While playing with some toy examples in the course of https://github.com/openjournals/joss-reviews/issues/766, I came up with the following questions I was curious about:

1. A common sub structure you see in binary formats is the length of some data encoded as a preceding integer. E.g. a string preceded by its length as a little endian 32-bit unsigned integer: `'\x11\x00\x00\x00hello world'`. How would one describe the second part of this data? I.e. a field whose size is that of a previously decoded field.
2. How would one describe alternatives for a field? The case I'm thinking about is decoding [CBOR](http://cbor.io) data. An integer has different encodings depending on its value and ideally I would like to describe "the next field is any one of the integer encodings." I don't immediately see how one would represent this for the purposes of *encoding* as opposed to decoding either.
3. How do you express more complex conditionals? E.g. if I have read three fields `a`, `b`, and `c`, how would I express a following field that is only present when all three of these are true (`a and b and c`)? I had some attempts at this but could only manage to trigger exceptions.
