### 2018-06-09 jfjlaros/bin-parser#13 @Smattr
inconsistent fmt quoting
**Describe the bug**
The examples sometimes quote the `fmt` parameter and sometimes do not. I gather in YAML either are permissible but I think to avoid confusion these should be made consistent.

**Additional context**
Related to https://github.com/openjournals/joss-reviews/issues/766.
