### 2018-06-09 jfjlaros/bin-parser#14 @Smattr
while documentation unclear
**Describe the bug**
The [while documentation](http://bin-parser.readthedocs.io/en/latest/structure.html#while-loops) shows an example of accessing the terminating loop element with `term`. However, it's not clear what the `list_term` value does in this scenario. Is the spec writer meant to use `list_term` later on to refer to this element? I don't follow from the docs whether `list_term` is the name of a variable or a directive to do something with the terminating value.

**Additional context**
Related to https://github.com/openjournals/joss-reviews/issues/766.
