### 2018-06-09 #766 @Smattr
@jfjlaros very nice project! The included presentation also gives a very nice motivation for how this project came about. I've posted a series of issues resulting from my review thus far. Some high level questions for you and perhaps @arfon as well:

1. Would you like us to submit nitpick feedback on coding style/conventions as well? E.g. more Pythonic idioms for doing certain things. If so, should these come as a separate issue or a PR or something else?
2. There seems to be very little error handling. I triggered multiple exceptions with malformed data or structure descriptions. This seems fine to me as this tool is a reversing aid not meant for end users, but perhaps JOSS has guidelines on this? From a personal standpoint, I think reversers generally prefer a bare exception trace as it helps you better debug your spec and incrementally reverse a format.
