### 2018-06-09 #766 @jfjlaros
Dear @Smattr,

Thank you for your kind words and your thorough review. I will try to address the issues you raised, and perhaps implement some of the features you requested. The latter will depend mainly on implementation time, but I think some can be addressed rather quickly.

As for detailed feedback, it is certainly welcome. I try to follow the recommended coding styles of the two programming languages used, but in this particular case I also try to keep the two code bases as similar as possible for maintainability reasons. This may result in some awkward code here and there. In any case, I would appreciate both general comments about my coding style and detailed remarks in any form you see fit.

I expected remarks about error handling and like you I am still on the fence on whether to implement it.
