### 2018-06-09 #766 @danielskatz
@Smattr - 1. Nitpicking items (that don't affect your checking boxes in the review, but would be nice to have) should probably be issues or PRs in the repo.
1. Re error handling - if you think this should block acceptance (ie you think the software really needs to have this to be useful), then let us know that - otherwise, again an issue in the repo that informs the developer but doesn't affect this submission would be good
