### 2018-06-09 #766 @Smattr
I think it's fine to leave as-is with regards to the error handling. I'll submit a PR with my nitpicks and we can discuss in more detail on that. Thanks, both.
