### 2018-06-09 jfjlaros/bin-parser#8 @Smattr
@jfjlaros if you're short on time, let me know and I can take a look at addressing this one as I already have a Python 3 environment setup. I think it's important to address this before submission because [Python 2 is due to be EOLed in 2020](https://pythonclock.org).
