### 2018-06-09 jfjlaros/bin-parser#15 @Smattr
Use of debug as a bit mask?
The debug value that gets propagated to `BinReader._debug` ends up being used in bit operations (e.g. [`if self._debug & 0x02:`](https://github.com/jfjlaros/bin-parser/blob/master/python/bin_parser.py#L365)). Meanwhile cli.py treats `debug` as a boolean. This leads to some unexpected behaviour. For example, `-d 4` on the command line results in less debugging information than `-d 3`. It looks to me like there are effectively three different debugging paths controlled by the same variable:

1. top level debugging output when `debug != 0`
2. some internal debugging when `debug & 0x1`
3. field debugging when `debug & 0x2`

Is this intentional?

Related to https://github.com/openjournals/joss-reviews/issues/766.
