### 2018-06-10 6f0705e
replace type(x) == type(y) with isinstance-testing
Using isinstance for type testing is a bit more Pythonic and allows easier
extension. For example, it's now possible to use a user-defined class in place
of dict as long as that class extends collections.Mapping.
