### 2018-06-10 75294a3
replace some maps with generators or list comprehensions
The use of comprehensions is more Pythonic and IMHO a bit clearer. This commit
has a couple of notable side effects:

  1. Previous calls to map that were relying on the Python 2 semantics
     (returning a list) have been turned into list comprehensions. This fixes a
     compatibility issue involving Python 3 changing the semantics of map to
     return a generator in a way in which the affected code previously was not
     anticipating.
  2. Calls to map that were not relying on the above-described Python 2
     semantics have been tu
[…]
