### 2018-06-10 3d6bc93
condense some code
