### 2018-06-10 jfjlaros/bin-parser#9 @jfjlaros
I think  771cfdf should solve this issue.
