### 2018-06-10 jfjlaros/bin-parser#10 @jfjlaros
This option was originally intended for a string of fixed length with a delimited string inside. In a later versions, this functionality is offered by using the `size` and `delimiter` parameters together. The only use left for `trim` is to control the padding character when writing such a field. I think I will remove this option.
