### 2018-06-10 cc603c8
Removed obsolete and undocumented `trim` parameter.
Fixes issue #10.
