### 2018-06-10 #766 @arfon
> @hainesr - I think so. @arfon, can you confirm?

Yes please.
