### 2018-06-10 d2ba920
Consistent quoting of `fmt` parameter.
Fixes issue #13.
