+++ examples/map/types.yml
@@ -4,7 +4,7 @@ types:
     function:
       name: struct
       args:
-        fmt: B
+        fmt: 'B'
         annotation:
           0x00: empty
           0xff: full
@@ -13,7 +13,7 @@ types:
     function:
       name: struct
       args:
-        fmt: <H
+        fmt: '<H'
         annotation:
           0x0000: zero
           0x0102: two hunderd and fifty-eight
