### 2018-06-10 jfjlaros/bin-parser#13 @jfjlaros
I agree.
