### 2018-06-10 jfjlaros/bin-parser#14 @jfjlaros
The terminating value is stored in this variable.

The non-terminating values are stored in the substructure, but the terminating value does not have an associated substructure, so we have to treat it separately.

Would an example help?
