### 2018-06-10 jfjlaros/bin-parser#15 @jfjlaros
Yes, this is intentional, but I can imagine it is a bit confusing. I'll add a guard for invalid debug levels and some documentation to explain the different levels.
