### 2018-06-10 jfjlaros/bin-parser#16 @hainesr
JS installation (from source) not working as documented?
**Describe the bug**
When installing the JS code from source a `bin_parser` executable isn't created in `./node_modules/.bin`.

**To Reproduce**
Steps to reproduce the behavior:
1. Follow the installation instructions ([here](http://bin-parser.readthedocs.io/en/latest/installation.html)) from source, up to and including `npm install .`

**Expected behavior**
I expected then to be able to run `./node_modules/.bin/bin-parser ...` as described in the next part of the documentation.

 - OS: Ubuntu Desktop 18.04
 - Version: 0.0.20

**Additional context**
I followed the installation instructions but I am not a JS expert and have a vanilla npm install provided by Ubuntu 18.04.

Related to openjournals/joss-reviews#766.
