### 2018-06-10 b02399a
Fix trivial doc typo: bin_parser -> bin-parser.
In the CLI section of the docs the example for the 'write' command with
the JS version incorrectly references 'bin_parser'.
