+++ docs/usage.rst
@@ -36,5 +36,5 @@ To convert a YAML file to binary, use the ``write`` subcommand:
 
 ::
 
-    ./node_modules/.bin/bin_parser write input.yml structure.yml types.yml \
+    ./node_modules/.bin/bin-parser write input.yml structure.yml types.yml \
       output.bin
