### 2018-06-10 jfjlaros/bin-parser#18 @hainesr
Please tag version 0.0.20 and create a GitHub release for JOSS submission.
For the JOSS submission please could you tag your version 0.0.20 in the code and create a corresponding "Release" in GitHub. See openjournals/joss-reviews#766 for the discussion of this. Thanks!
