### 2018-06-10 #766 @hainesr
Thanks @arfon. I have raised an issue.
