### 2018-06-10 c55f9a3
More information about debugging options.
Added a section about debugging in the documentation and added a check for
invalid debugging options.

Fixes #15.
