### 2018-06-10 e4dda72
Added documentation for running JavaScript CLI from source.
Fixes #16.
