+++ docs/usage.rst
@@ -38,3 +38,9 @@ To convert a YAML file to binary, use the ``write`` subcommand:
 
     ./node_modules/.bin/bin_parser write input.yml structure.yml types.yml \
       output.bin
+
+Please note that when installing from source, the ``bin_parser`` executable is
+not installed. Instead run the script ``cli.js`` as follows:
+
+::
+    nodejs javascript/cli.js
