### 2018-06-10 jfjlaros/bin-parser#16 @jfjlaros
It looks like `npm` behaves differently when installing from source. This may be a bug in `npm`, but I am not sure, I am not a JavaScript expert myself.

I added a line in the installation instructions to address this.
