### 2018-06-10 jfjlaros/bin-parser#18 @jfjlaros
I was planning to do so after I address the comments. Would that be okay?
