### 2018-06-10 jfjlaros/bin-parser#18 @hainesr
Yes, I think that's the best course of action. Thanks!
