### 2018-06-10 4c73c5b
Added size_string example.
