### 2018-06-10 jfjlaros/bin-parser#12 @jfjlaros
For question 1: I made an example in 4c73c5b.
