### 2018-06-10 9e26cb0
Added functionality for variable types.
- The `type` field can now be a variable name.
- Added an example.

This should cover question 2 of issue #12.
