### 2018-06-10 jfjlaros/bin-parser#12 @jfjlaros
For question 2: I never considered variable types before, but it was not difficult to implement. See 9e26cb0 for an example.
