### 2018-06-10 jfjlaros/bin-parser#12 @jfjlaros
For question 3: This is currently not supported.

Support for more operands like in your example would be easy, but support for arbitrary expressions will require some work. I'll have to think about how to implement this.
