### 2018-06-10 jfjlaros/bin-parser#8 @jfjlaros
If it would not be too much trouble, yes please.
