### 2018-06-11 jfjlaros/bin-parser#12 @jfjlaros
I think it would be best to add support for nested structures here, perhaps something along these lines:

```yaml
if:
  operands:
    - operands:
        - 1
        - 1
      operator: eq
    - operands:
        - 2
        - 3
      operator: ne
  operator: and
```
