### 2018-06-11 #766 @jfjlaros
Dear @hainesr,

Thank you very much for your suggestions and fixes.
