### 2018-06-11 4f76674
Added recursive evaluation of expressions.
