### 2018-06-11 jfjlaros/bin-parser#12 @jfjlaros
This seems to be working, see 4f76674 for an example.
