### 2018-06-11 jfjlaros/bin-parser#18 @jfjlaros
I do have to increase the version number to test npm and PyPI. Do you think that would that be a problem?
