### 2018-06-11 016c48f
Synchronised JavaScript and Python executable names.
