### 2018-06-11 5f400b8
Added tests.
- Unit tests for the new functionalities listed in #12.
- Comparison tests for all examples.
