+++ tests/test_parser.py
@@ -29,7 +29,7 @@ def _idempotence(path, input_file, structure_file, types_file):
 
 class TestReader(object):
     """
-    Test the python.fam_parser module.
+    Test the python.bin_parser module.
     """
     def setup(self):
         self._data = {
@@ -47,7 +47,20 @@ class TestReader(object):
                 'var_size', 'var_size.dat', 'structure.yml', 'types.yml'],
             'padding': [
                 'padding', 'padding.dat', 'structure.yml', 'types.yml'],
-            'order': ['order', 'order.dat', 'structure.yml', 'types.yml']}
+            'order': ['order', 'order.dat', 'structure.yml', 'types.yml'],
+
+            'colour': ['colour', 'colour.dat', 'structure.yml', 'types.yml'],
+            'complex_eval': [
+                'complex_eval', 'complex_eval.dat', 'structure.yml',
+                'types.yml'],
+            'csv': ['csv', 'test.csv', 'structure.yml', 'types.yml'],
+            'flags': ['flags', 'flags.dat', 'structure.yml', 'types.yml'],
+            'map': ['map', 'map.dat', 'structure.yml', 'types.yml'],
+            'size_string': [
+                'size_string', 'size_string.dat', 'structure.yml',
+                'types.yml'],
+            'var_type': [
+                'var_type', 'var_type.dat', 'structure.yml', 'types.yml']}
 
     def test_idempotence(self):
         for example in self._data:
@@ -125,3 +138,63 @@ class TestReader(object):
 
     def test_order_2(self):
         assert _bin_reader(*self._data['order'])['val_2'] == 0x010200
+
+    def test_colour_1(self):
+        assert _bin_reader(*self._data['colour'])['background']['g'] == 128
+
+    def test_colour_2(self):
+        assert _bin_reader(*self._data['colour'])['background']['b'] == 'full'
+
+    def test_complex_eval_1(self):
+        assert 'item_1' not in _bin_reader(*self._data['complex_eval'])
+
+    def test_complex_eval_2(self):
+        assert _bin_reader(*self._data['complex_eval'])['item_2'] == 'A'
+
+    def test_flags_1(self):
+        assert not _bin_reader(*self._data['flags'])['flags']['bit_one']
+
+    def test_flags_2(self):
+        assert _bin_reader(*self._data['flags'])['flags']['bit_two']
+
+    def test_flags_3(self):
+        assert _bin_reader(*self._data['flags'])['flags']['flag_04']
+
+    def test_map_1(self):
+        assert _bin_reader(
+            *self._data['map'])['number'] == 'two hunderd and fifty-eight'
+
+    def test_map_2(self):
+        assert _bin_reader(*self._data['map'])['choice'] == 1
+
+    def test_size_string_1(self):
+        assert _bin_reader(
+            *self._data['size_string'])['string_1']['size_of_string'] == 28
+
+    def test_size_string_2(self):
+        assert len(_bin_reader(
+            *self._data['size_string'])['string_1']['string']) == 28
+
+    def test_size_string_3(self):
+        assert _bin_reader(
+            *self._data['size_string'])['string_2']['size_of_string'] == 23
+
+    def test_size_string_4(self):
+        assert len(_bin_reader(
+            *self._data['size_string'])['string_2']['string']) == 23
+
+    def test_var_type_1(self):
+        assert _bin_reader(
+            *self._data['var_type'])['value_1']['content'] == 'A'
+
+    def test_var_type_2(self):
+        assert _bin_reader(
+            *self._data['var_type'])['value_1']['type_name'] == 'char'
+
+    def test_var_type_3(self):
+        assert _bin_reader(
[…]
