### 2018-06-11 a69ced5
Some minor formatting.
