### 2018-06-11 jfjlaros/bin-parser#18 @hainesr
I presume you can then up the version number for the JOSS submission. Is that right @danielskatz, @arfon
?
