### 2018-06-11 #766 @hainesr
@danielskatz, @arfon - There is a question about whether the version pointed to by this submission can change now there have been lots of updates to this project here: jfjlaros/bin-parser#18.
