### 2018-06-11 #766 @danielskatz
yes, it's fine to change the version, in both the paper.md and the repo
