### 2018-06-12 jfjlaros/bin-parser#18 @arfon
> I presume you can then up the version number for the JOSS submission. Is that right @danielskatz, @arfon

Yep.
