### 2018-06-12 jfjlaros/bin-parser#12 @Smattr
Ah nice! I didn't realise many of these had such simple solutions! For (3), a future extension could be to let people more naturally express these using Python syntax. I.e. just write something like:

```
if:
  - '1 == 1 and 2 != 3'
```

and then decode this with something like [Python ast](https://docs.python.org/3/library/ast.html). Looks pretty good for now though.
