### 2018-06-12 jfjlaros/bin-parser#14 @Smattr
I think I follow what you're saying now, but yes an example would clarify this.
