### 2018-06-12 4c51623
Added `while` example.
Fixes #14.
