### 2018-06-12 jfjlaros/bin-parser#14 @jfjlaros
I hope the example in 4c51623 helps.
