### 2018-06-12 4717f7a
Updated documentation.
- Fields of which the length depends on a variable.
- Variable types.
- Complex conditionals.

This should finish #12.
