### 2018-06-13 7554822
adjust extras/test.sh shebang for wider compatibility
On some Linux and BSD distributions Bash is not installed at /bin/bash but
rather somewhere like /usr/local/bin/bash. Using /usr/bin/env in this shell
script's shebang lets us find it wherever it is.
