### 2018-06-13 jfjlaros/bin-parser#14 @Smattr
Ah yes! This both clarifies the situation and it also was *not* what I had previously thought. I had previously thought `list_term` was intended to be used as a variable somewhere later in your structure YAML but I see now that it simple appears as a label in your output. Thanks!
