### 2018-06-13 9f028dc
Python 3 compatibility.
- Fixed conversion bytes <-> unicode and vice versa in the library.
- Use binary read and write mode in CLI when needed.
- Fixed print statements in examples.

TODO:
- Check `make_skeleton.py`.
- Update documentation.

Affects issue #8.
