+++ tests/test_functions.py
@@ -22,70 +22,74 @@ class TestParser(object):
             getattr(self._brf, func)(data, **kwargs), **kwargs) == data
 
     def test_raw_single(self):
-        assert self._brf.raw('\x03') == '03'
+        assert self._brf.raw(b'\x03') == '03'
 
     def test_raw_multi(self):
-        assert self._brf.raw('\x00\x01\x02') == '00 01 02'
+        assert self._brf.raw(b'\x00\x01\x02') == '00 01 02'
 
     def test_raw_idem(self):
-        self._idem('raw', '\x00\x01\x02')
+        self._idem('raw', b'\x00\x01\x02')
 
     def test_bit(self):
-        assert self._brf.bit('\x03') == '00000011'
+        assert self._brf.bit(b'\x03') == '00000011'
 
     def test_bit_idem(self):
-        self._idem('bit', '\x03')
+        self._idem('bit', b'\x03')
 
     def test_le_s_short(self):
-        assert self._brf.struct('\x01\x00', fmt='<h') == 1
+        assert self._brf.struct(b'\x01\x00', fmt='<h') == 1
 
     def test_be_s_short(self):
-        assert self._brf.struct('\x00\x01', fmt='>h') == 1
+        assert self._brf.struct(b'\x00\x01', fmt='>h') == 1
 
     def test_le_s_short_idem(self):
-        self._idem('struct', '\x01\x02', fmt='<h')
+        self._idem('struct', b'\x01\x02', fmt='<h')
 
     def test_be_s_float_idem(self):
-        self._idem('struct', '\x01\x02\x03\x04', fmt='>f')
+        self._idem('struct', b'\x01\x02\x03\x04', fmt='>f')
 
     def test_labels(self):
         assert self._brf.struct(
-            '\x01\x02', fmt='BB', labels=['a', 'b']) == {'a': 1, 'b': 2}
+            b'\x01\x02', fmt='BB', labels=['a', 'b']) == {'a': 1, 'b': 2}
 
     def test_labels_idem(self):
-        self._idem('struct', '\x01\x02', fmt='BB', labels=['a', 'b'])
+        self._idem('struct', b'\x01\x02', fmt='BB', labels=['a', 'b'])
 
     def test_annotation(self):
         assert self._brf.struct(
-            '\x01\x02', fmt='BB', annotation={1: 'x'}) == ['x', 2]
+            b'\x01\x02', fmt='BB', annotation={1: 'x'}) == ['x', 2]
 
     def test_annotation_idem(self):
-        self._idem('struct', '\x01\x02', fmt='BB', annotation={1: 'a'})
+        self._idem('struct', b'\x01\x02', fmt='BB', annotation={1: 'a'})
 
     def test_labels_annotation(self):
         assert self._brf.struct(
-            '\x01\x02', fmt='BB', labels=['a', 'b'],
+            b'\x01\x02', fmt='BB', labels=['a', 'b'],
             annotation={1: 'x'}) == {'a': 'x', 'b': 2}
 
     def test_labels_annotation_idem(self):
         self._idem(
-            'struct', '\x01\x02', fmt='BB', labels=['a', 'b'],
+            'struct', b'\x01\x02', fmt='BB', labels=['a', 'b'],
             annotation={1: 'x'})
 
     def test_flags(self):
-        assert self._brf.flags('\x03', {}) == {'flag_01': True, 'flag_02': True}
+        assert self._brf.flags(b'\x03', {}) == {
+            'flag_01': True, 'flag_02': True}
 
     def test_flags_false(self):
-        assert self._bwf.flags({'flag_01': True, 'flag_02': False}, {}) == '\x01'
+        assert self._bwf.flags(
[…]
