### 2018-06-13 jfjlaros/bin-parser#8 @jfjlaros
@Smattr I hope you didn't start on the Python 3 compatibility issue yet, I started on this in 9f028dc.
