### 2018-06-14 jfjlaros/bin-parser#8 @Smattr
Ah nice! I had started, but was having a bit of difficulty figuring out which variables were meant to be strings and which were byte data. It looks like your changes subsume what I have. I'll test the latest version with Python 3 and report back.
