### 2018-06-14 b039858
Last changes for Python 3 compatibility.
