+++ extras/make_skeleton.py
@@ -30,7 +30,8 @@ def make_structure(data, delimiter):
     """
     structure = []
 
-    for index in range(len(data.split(''.join(map(chr, delimiter))))):
+    for index in range(
+            len(data.split(''.join(map(chr, delimiter)).encode('utf-8')))):
         structure.append({
             'name': 'field_{:06d}'.format(index),
             'type': 'raw'})
@@ -42,11 +43,11 @@ def make_types(delimiter):
     return {
         'types': {
             'raw': {
-                'delimiter': map(HexInt, delimiter),
+                'delimiter': list(map(HexInt, delimiter)),
                 'function': {
                     'name': 'raw'}},
             'text': {
-                'delimiter': map(HexInt, delimiter)}}}
+                'delimiter': list(map(HexInt, delimiter))}}}
 
 
 def make_skeleton(input_handle, structure_handle, types_handle, delimiter):
@@ -78,7 +79,7 @@ def main():
         formatter_class=argparse.RawDescriptionHelpFormatter)
 
     parser.add_argument(
-        'input_handle', metavar='INPUT', type=argparse.FileType('r'),
+        'input_handle', metavar='INPUT', type=argparse.FileType('rb'),
         help='input file')
     parser.add_argument(
         'structure_handle', metavar='STRUCTURE', type=argparse.FileType('w'),
