### 2018-06-14 jfjlaros/bin-parser#8 @jfjlaros
@Smattr In b039858 I added changes to the scripts in "extras" and the documentation. All tests pass for Python 2.7 and Python 3.5 on my machine.
