### 2018-06-15 jfjlaros/bin-parser#8 @Smattr
Looks great! Things seem to run well under Python 3 for me now.

One issue: tests/test_parser.py defines the method `TestReader.test_var_type_2` twice. I picked this up by running one of my goto sanity checks from the root directory:

```bash
$ find . -iname "*.py" -exec echo "{}" \; -exec python -m pylint --errors-only "{}" \;
...
./tests/test_parser.py
E:199, 4: method already defined line 191 (function-redefined)
...
```
