### 2018-06-15 #766 @Smattr
At this point, I'm happy with the submission after the Python 3 compatibility branch is merged (https://github.com/jfjlaros/bin-parser/issues/8) and we have a tagged version (https://github.com/jfjlaros/bin-parser/issues/18). Well done addressing everything so fast, @jfjlaros.
