### 2018-06-15 13f0842
Fixed already defined method bug.
Fixes #8.
