### 2018-06-15 jfjlaros/bin-parser#18 @jfjlaros
I just released version 1.0.0.
