### 2018-06-15 #766 @jfjlaros
I released [v1.0.0](https://github.com/jfjlaros/bin-parser/releases/tag/v1.0.0).

Thanks @Smattr, @hainesr for your suggestions and contributions. It really improved the quality of the package.
