+++ docs/credits.rst
@@ -1,0 +1,14 @@
+Contributors
+============
+
+- Jeroen F.J. Laros <J.F.J.Laros@lumc.nl> (Original author, maintainer)
+- Daniel S. Katz <d.katz@ieee.org>
+- Jamie Ross <jamie.ross@electricityexchange.ie>
+- Matthew Fernandez <matthew.fernandez@gmail.com>
+- Robert Haines <rhaines@manchester.ac.uk>
+
+Find out who contributed:
+
+::
+
+    git shortlog -s -e
