### 2018-06-16 #766 @jfjlaros
I added credits in c1f2e6e, do we also want these in the release?
