### 2018-06-20 8abc300
Added Travis build satus.
