+++ README.rst
@@ -1,6 +1,9 @@
 General binary file parser
 ==========================
 
+.. image:: https://travis-ci.org/jfjlaros/bin-parser.svg?branch=master
+   :target: https://travis-ci.org/jfjlaros/bin-parser
+
 This library provides general binary file parsing by interpreting documentation
 of a file structure and data types. By default, it supports basic data types
 like big-endian and little-endian integers, floats and doubles, variable length
