+++ .travis.yml
@@ -1,1 +1,6 @@
 language: python
+python:
+    - "2.7"
+    - "3.5"
+install: pip install tox-travis
+script: tox
