+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -37,4 +37,4 @@ Provide any other important details below.
 :question:
 
 [1]: https://github.com/jfjlaros/bin-parser/pulls
-[2]: https://github.com/jfjlaros/bin-parser/blob/master/.github/CONTRIBUTING.md#code-style
+[2]: https://github.com/jfjlaros/bin-parser/blob/master/docs/CONTRIBUTING.md#code-style
