### 2018-06-20 7d49ac8
Version and license badges.
