+++ README.md
@@ -3,6 +3,9 @@ General binary file parser
 
 [![Build Status](https://travis-ci.org/jfjlaros/bin-parser.svg?branch=master)](https://travis-ci.org/jfjlaros/bin-parser)
 [![Docs Status](https://readthedocs.org/projects/bin-parser/badge/?version=latest)](https://bin-parser.readthedocs.io/en/latest)
+[![PyPI Version](https://img.shields.io/pypi/v/bin-parser.svg)](https://pypi.org/project/bin-parser/)
+[![NPM Version](https://img.shields.io/npm/v/bin-parser.svg)](https://www.npmjs.com/package/bin-parser)
+[![License](https://img.shields.io/pypi/l/bin-parser.svg)](https://raw.githubusercontent.com/jfjlaros/bin-parser/master/LICENSE.md)
 
 ---
 
