### 2018-06-20 a951c5e
Updated documentation.
