+++ README.rst
@@ -9,6 +9,8 @@ General binary file parser
    :target: https://pypi.org/project/bin-parser/
 .. image:: https://img.shields.io/npm/v/bin-parser.svg
    :target: https://www.npmjs.com/package/bin-parser
+.. image:: https://img.shields.io/github/release/jfjlaros/bin-parser.svg
+   :target: https://github.com/jfjlaros/bin-parser/releases
 .. image:: https://img.shields.io/pypi/l/bin-parser.svg
    :target: https://raw.githubusercontent.com/jfjlaros/bin-parser/master/LICENSE.md
 
