### 2018-06-20 eb5d5f3
Updated documentation.
