### 2018-06-21 0aa4945
Updated documentation.
