### 2018-06-21 #766 @danielskatz
@jfjlaros , I think that's up to you.

Let me know if you think this is ready to publish, or if you want to make any final changes.
