### 2018-06-21 #766 @jfjlaros
@danielskatz, I added the credits and automatic testing with Travis to [v1.0.1](https://github.com/jfjlaros/bin-parser/releases/tag/v1.0.1). I think that wraps it up.
