### 2018-06-21 #766 @danielskatz
Is 1.0.1 archived somewhere with a DOI?  If not, please do this.

And then paste the DOI here.
