+++ README.md
@@ -13,6 +13,7 @@ General binary file parser
 [![image](https://img.shields.io/github/languages/top/jfjlaros/bin-parser.svg)](https://github.com/jfjlaros/bin-parser)
 [![image](https://img.shields.io/github/license/jfjlaros/bin-parser.svg)](https://raw.githubusercontent.com/jfjlaros/bin-parser/master/LICENSE.md)
 [![image](https://zenodo.org/badge/DOI/10.5281/zenodo.1295625.svg)](https://zenodo.org/record/1295625)
+[![image](http://joss.theoj.org/papers/345778df6874891e84c712cc0d665549/status.svg)](http://joss.theoj.org/papers/345778df6874891e84c712cc0d665549)
 
 ------------------------------------------------------------------------
 
