+++ README.rst
@@ -25,6 +25,8 @@ General binary file parser
    :target: https://raw.githubusercontent.com/jfjlaros/bin-parser/master/LICENSE.md
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.1295625.svg
    :target: https://zenodo.org/record/1295625
+.. image:: http://joss.theoj.org/papers/345778df6874891e84c712cc0d665549/status.svg
+   :target: http://joss.theoj.org/papers/345778df6874891e84c712cc0d665549
 
 ----
 
