### 2018-06-21 #766 @danielskatz
👋  @arfon - over to you for final acceptance

@Smattr, @hainesr - Thanks for your reviewing and contributions!
