### 2018-06-21 #766 @jfjlaros
@arfon, @danielskatz: Thank you for editing this submission.

@Smattr, @hainesr: Thank you for your suggestions and contributions. I really appreciate it.
