### 2018-06-22 #766 @Smattr
This was the first JOSS review I've been involved with and I thought the whole thing went very well. Thanks all :)
